+++ src/hooks/useDocumentTitle.js
@@ -1,0 +1,9 @@
+import React, { useEffect } from "react";
+
+const useDocumentTitle = (title) => {
+    useEffect(() => {
+        document.title = title
+    })
+}
+
+export default useDocumentTitle;
